generalize package
